+++ src/components/InlineLoader/interfaces.ts
@@ -1,0 +1,4 @@
+import { HTMLAttributes } from 'react';
+export interface IInlineLoaderProps extends HTMLAttributes<HTMLDivElement> {
+  error?: boolean;
+}