+++ src/components/Input/interfaces.ts
@@ -1,7 +1,8 @@
-import { InputHTMLAttributes } from 'react';
+import { InputHTMLAttributes, ReactNode } from 'react';
 
 export interface IInput extends InputHTMLAttributes<HTMLInputElement> {
   error?: boolean | string;
   success?: boolean;
   disabled?: boolean;
+  children?: ReactNode;
 }