+++ src/components/Modal/index.tsx
@@ -1,15 +1,17 @@
-import { FC, HTMLProps, ReactNode } from 'react';
+import { FC, HTMLAttributes, ReactNode } from 'react';
 import * as S from './styles';
 
 export type ModalProps = {
   show: boolean;
   children: ReactNode;
-} & HTMLProps<HTMLDivElement>;
+} & HTMLAttributes<HTMLDivElement>;
 
 const Modal: FC<ModalProps> = ({ show, children, className, ...props }) =>
   show ? (
     <S.Overlay>
-      <S.Container className={className}>{children}</S.Container>
+      <S.Container className={className} {...props}>
+        {children}
+      </S.Container>
     </S.Overlay>
   ) : null;
 export default Modal;